fix constant of shape type

+++ onnx2kerastl/constant_layers.py
@@ -20,7 +20,7 @@ def convert_constant_of_shape(node, params, layers, lambda_func, node_name, kera
     value = params.get('value')
     if value is None:
         raise NotImplementedError("ConstantOfShape should have a value param")
-    layers[node_name] = np.ones(layers[node.input[0]], dtype=np.int32)*params['value']
+    layers[node_name] = np.ones(layers[node.input[0]], dtype=value.dtype)*params['value']
 
 
 def convert_one_hot(node, params, layers, lambda_func, node_name, keras_name):
